cardDefinition "refactor"
