+++ src/domain/cardImage.ts
@@ -1,0 +1,7 @@
+import { MergeParameters } from "./mergeParameters.js"
+
+export interface CardImage {
+    buffer?: Buffer
+    merge?: MergeParameters
+    create: () => Promise<void>
+}
