+++ src/domain/cardType.ts
@@ -1,0 +1,4 @@
+import { z } from "zod"
+
+export const cardTypeSchema = z.enum(["melee", "range", "mass"])
+export type CardType = z.infer<typeof cardTypeSchema>
