+++ src/domain/mergeParameters.ts
@@ -1,0 +1,10 @@
+import { BlendMode } from "jimp"
+
+export interface MergeParameters {
+    x: number
+    y: number
+    blend?: {
+        mode: BlendMode,
+        opacitySource: number
+    }
+}
